+++ src/main/java/miyucomics/hexical/casting/patterns/conjure/OpConjureSpike.kt
@@ -7,6 +7,7 @@ import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.getBlockPos
 import at.petrak.hexcasting.api.casting.getPositiveDoubleUnderInclusive
 import at.petrak.hexcasting.api.casting.iota.Iota
+import at.petrak.hexcasting.api.casting.mishaps.MishapBadBlock
 import at.petrak.hexcasting.api.casting.mishaps.MishapInvalidIota
 import at.petrak.hexcasting.api.misc.MediaConstants
 import miyucomics.hexical.entities.SpikeEntity
@@ -23,10 +24,10 @@ class OpConjureSpike : SpellAction {
 	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
 		val position = args.getBlockPos(0, argc)
 		env.assertPosInRange(position)
-		if (env.world.getBlockState(position).isAir)
-			throw MishapInvalidIota.of(args[0], 2, "solid_block")
 		val offset = args.getBlockPos(1, argc)
 		val direction = Direction.fromVector(offset.x, offset.y, offset.z) ?: throw MishapInvalidIota.of(args[1], 1, "axis_vector")
+		if (!env.world.getBlockState(position).isSideSolidFullSquare(env.world, position, direction))
+			throw MishapBadBlock.of(position, "solid_platform")
 		if (env.world.getEntitiesByType(HexicalEntities.SPIKE_ENTITY, Box.of(Vec3d.ofCenter(position.add(offset)), 0.9, 0.9, 0.9)) { true }.size > 0)
 			return SpellAction.Result(Noop(position), 0, listOf())
 		val delay = floor(args.getPositiveDoubleUnderInclusive(2, 10.0, argc) * 20.0).toInt()